+++ packages/aws-cdk/lib/cli/telemetry/endpoint-client.ts
@@ -5,7 +5,7 @@ import type { UrlWithStringQuery } from 'url';
 import { ToolkitError } from '@aws-cdk/toolkit-lib';
 import { IoHelper } from '../../api-private';
 import type { IIoHost } from '../io-host';
-import type { ITelemetrySink } from './client-interface';
+import type { ITelemetrySink } from './sink-interface';
 import type { TelemetrySchema } from './schema';
 
 const REQUEST_ATTEMPT_TIMEOUT_MS = 2_000;
