+++ packages/aws-cdk/lib/cli/telemetry/io-host-client.ts
@@ -1,5 +1,5 @@
 import type { IIoHost } from '@aws-cdk/toolkit-lib';
-import type { ITelemetrySink } from './client-interface';
+import type { ITelemetrySink } from './sink-interface';
 import type { TelemetrySchema } from './schema';
 import { IoHelper } from '../../api-private';
